i2s

+++ src/esp/ringbuf.h
@@ -59,8 +59,11 @@ void rb_reset_and_abort_write(ringbuf_t *rb);
 void rb_stat(ringbuf_t *rb);
 ssize_t rb_filled(ringbuf_t *rb);
 ssize_t rb_available(ringbuf_t *rb);
-int rb_read(ringbuf_t *rb, uint8_t *buf, int len, uint32_t ticks_to_wait);
-int rb_write(ringbuf_t *rb, const uint8_t *buf, int len,
+//int rb_read(ringbuf_t *rb, uint8_t *buf, int len, uint32_t ticks_to_wait);
+int rb_read(ringbuf_t *rb, int16_t *buf, int len, uint32_t ticks_to_wait);
+//int rb_write(ringbuf_t *rb, const uint8_t *buf, int len,
+//             uint32_t ticks_to_wait);
+int rb_write(ringbuf_t *rb, const int16_t *buf, int len,
              uint32_t ticks_to_wait);
 void rb_cleanup(ringbuf_t *rb);
 void rb_signal_writer_finished(ringbuf_t *rb);
